+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added at [0.2.3]
 
+- **File Sync:** The 'Attached Files' panel now automatically updates when files are changed externally.
+
 ### Changed at [0.2.3]
 
 ### Fixed at [0.2.3]